+++ cropper/src/main/java/com/canhub/cropper/CropImageOptions.kt
@@ -9,9 +9,12 @@ import android.os.Parcel
 import android.os.Parcelable
 import android.text.TextUtils
 import android.util.TypedValue
+import androidx.annotation.ColorInt
 import com.canhub.cropper.CropImageView.CropShape
 import com.canhub.cropper.CropImageView.Guidelines
 import com.canhub.cropper.CropImageView.RequestSizeOptions
+
+private val COLOR_PURPLE = Color.rgb(153, 51, 153)
 
 /**
  * All the possible options that can be set to customize crop image.<br></br>
@@ -96,6 +99,11 @@ open class CropImageOptions : Parcelable {
     @JvmField
     var showProgressBar: Boolean
 
+    /** The color of the progress bar. Only works on API level 21 and upwards. */
+    @JvmField
+    @ColorInt
+    var progressBarColor: Int
+
     /**
      * if auto-zoom functionality is enabled.<br></br>
      * default: true.
@@ -140,6 +148,7 @@ open class CropImageOptions : Parcelable {
 
     /** the color of the guidelines lines  */
     @JvmField
+    @ColorInt
     var borderLineColor: Int
 
     /** thickness of the corner line. (in pixels)  */
@@ -156,6 +165,7 @@ open class CropImageOptions : Parcelable {
 
     /** the color of the corner line  */
     @JvmField
+    @ColorInt
     var borderCornerColor: Int
     /**
      * The fill color of circle corner
@@ -169,6 +179,7 @@ open class CropImageOptions : Parcelable {
 
     /** the color of the guidelines lines  */
     @JvmField
+    @ColorInt
     var guidelinesColor: Int
 
     /**
@@ -176,6 +187,7 @@ open class CropImageOptions : Parcelable {
      * crop window.
      */
     @JvmField
+    @ColorInt
     var backgroundColor: Int
 
     /** the min width the crop window is allowed to be. (in pixels)  */
@@ -220,6 +232,7 @@ open class CropImageOptions : Parcelable {
 
     /** the color to use for action bar items icons  */
     @JvmField
+    @ColorInt
     var activityMenuIconColor: Int
 
     /** the Android Uri to save the cropped image to  */
@@ -323,6 +336,7 @@ open class CropImageOptions : Parcelable {
 
     /** The default cropper label text color **/
     @JvmField
+    @ColorInt
     var cropperLabelTextColor: Int
 
     /** The default cropper label text **/
@@ -344,6 +358,7 @@ open class CropImageOptions : Parcelable {
         scaleType = CropImageView.ScaleType.FIT_CENTER
         showCropOverlay = true
         showProgressBar = true
+        progressBarColor = COLOR_PURPLE
         autoZoomEnabled = true
         multiTouchEnabled = false
         centerMoveEnabled = true
@@ -409,6 +424,7 @@ open class CropImageOptions : Parcelable {
         scaleType = CropImageView.ScaleType.values()[parcel.readInt()]
         showCropOverlay = parcel.readByte().toInt() != 0
         showProgressBar = parcel.readByte().toInt() != 0
+        progressBarColor = parcel.readInt()
         autoZoomEnabled = parcel.readByte().toInt() != 0
         multiTouchEnabled = parcel.readByte().toInt() != 0
         centerMoveEnabled = parcel.readByte().toInt() != 0
@@ -474,6 +490,7 @@ open class CropImageOptions : Parcelable {
         dest.writeInt(scaleType.ordinal)
         dest.writeByte((if (showCropOverlay) 1 else 0).toByte())
         dest.writeByte((if (showProgressBar) 1 else 0).toByte())
+        dest.writeInt(progressBarColor)
         dest.writeByte((if (autoZoomEnabled) 1 else 0).toByte())
         dest.writeByte((if (multiTouchEnabled) 1 else 0).toByte())
         dest.writeByte((if (centerMoveEnabled) 1 else 0).toByte())