+++ cropper/src/main/java/com/canhub/cropper/CropImageView.kt
@@ -2,6 +2,7 @@ package com.canhub.cropper
 
 import android.app.Activity
 import android.content.Context
+import android.content.res.ColorStateList
 import android.graphics.Bitmap
 import android.graphics.Bitmap.CompressFormat
 import android.graphics.BitmapFactory
@@ -10,6 +11,7 @@ import android.graphics.Matrix
 import android.graphics.Rect
 import android.graphics.RectF
 import android.net.Uri
+import android.os.Build
 import android.os.Bundle
 import android.os.Parcelable
 import android.provider.MediaStore
@@ -1962,6 +1964,9 @@ class CropImageView @JvmOverloads constructor(context: Context, attrs: Attribute
         mCropOverlayView.setCropWindowChangeListener(this)
         mCropOverlayView.setInitialAttributeValues(options)
         mProgressBar = v.findViewById(R.id.CropProgressBar)
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
+            mProgressBar.indeterminateTintList = ColorStateList.valueOf(options.progressBarColor)
+        }
         setProgressBarVisibility()
     }